Move item Status to its own module
Re #310
